+++ apps/workflows-dashboard/src/pages/Overview/hooks/useCaseResolvingStatsQuery/useCaseResolvingStatsQuery.ts
@@ -1,13 +1,14 @@
 import { userStatsQueryKeys } from '@app/domains/user/api/user-stats';
 import { useQuery } from '@tanstack/react-query';
-import * as dayjs from 'dayjs';
+import dayjs from 'dayjs';
 import { useMemo } from 'react';
 
 export const useCaseResolvingStatsQuery = () => {
-  const initialDate = useMemo(() => +dayjs().subtract(6, 'days').startOf('day').toDate(), []);
-  const { data = [], isLoading } = useQuery(
-    userStatsQueryKeys.userCaseResolvingStats({ fromDate: initialDate }),
-  );
+  const initialDate = useMemo(() => +dayjs().subtract(6, 'day').startOf('day').toDate(), []);
+  const { data = [], isLoading } = useQuery({
+    ...userStatsQueryKeys.userDailyCasesResolvedStats({ fromDate: initialDate }),
+    enabled: Boolean(initialDate),
+  });
 
   return {
     data,