+++ services/workflows-service/scripts/seed.ts
@@ -778,7 +778,7 @@ async function seed(bcryptSalt: Salt) {
       },
     },
     where: {
-      workflowDefinitionId: manualMachineId,
+      workflowDefinitionId: riskScoreMachineKybId,
       endUserId: { not: null },
     },
   });